+++ apps/block_scout_web/test/block_scout_web/controllers/api/rpc/token_controller_test.exs
@@ -85,25 +85,25 @@ defmodule BlockScoutWeb.API.RPC.TokenControllerTest do
     end
   end
 
-  defp gettoken_schema do
-    ExJsonSchema.Schema.resolve(%{
-      "type" => "object",
-      "properties" => %{
-        "message" => %{"type" => "string"},
-        "status" => %{"type" => "string"},
-        "result" => %{
-          "type" => "object",
-          "properties" => %{
-            "name" => %{"type" => "string"},
-            "symbol" => %{"type" => "string"},
-            "totalSupply" => %{"type" => "string"},
-            "decimals" => %{"type" => "string"},
-            "type" => %{"type" => "string"},
-            "cataloged" => %{"type" => "string"},
-            "contractAddress" => %{"type" => "string"}
-          }
-        }
-      }
-    })
-  end
+  # defp gettoken_schema do
+  #   ExJsonSchema.Schema.resolve(%{
+  #     "type" => "object",
+  #     "properties" => %{
+  #       "message" => %{"type" => "string"},
+  #       "status" => %{"type" => "string"},
+  #       "result" => %{
+  #         "type" => "object",
+  #         "properties" => %{
+  #           "name" => %{"type" => "string"},
+  #           "symbol" => %{"type" => "string"},
+  #           "totalSupply" => %{"type" => "string"},
+  #           "decimals" => %{"type" => "string"},
+  #           "type" => %{"type" => "string"},
+  #           "cataloged" => %{"type" => "string"},
+  #           "contractAddress" => %{"type" => "string"}
+  #         }
+  #       }
+  #     }
+  #   })
+  # end
 end